working loading and saving of files with dialog from rust
This means we are loading and saving files with the rfd crate which
utilizes platform appropriate dialogs for saving and loading files. We
can start to use this in all other file dialog locations to make sure
we are getting the right files.

It uses xdg-desktop-portals on linux so we will always use the
appropriate file dialog for each desktop environment.