adding song_lyrics.rs
This module will allow for searching and finding songs in genius
lyrics and giving me the ability to get the lyrics from those songs.

The main function implemented is search_song first. From here we can
get the id of the songs we are looking for and then add the lyrics
into Lumina.