adding a way for items to propagate to main app layer
If I am not mistaken, I may not need to do this.
There might be a way to make items appear by using the DndSource and
DndDrop modules. I hope that these essentially move the item to a
global level and then drop the data in some way to the nav_bar.